requirements.txt
discord.py
coloredlogs
crontab